# openAPI.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LocalDNS API",
    "version": "1.0.0",
    "description": "Backend API for LocalDNS"
  },
  "servers": [
    {
      "url": "http://localhost:4322"
    }
  ],
  "components": {
    "schemas": {
...
      "CreateDomainRequest": {
        "type": "object",
        "properties": {
          "domain": {
            "type": "string",
            "example": "api.localhost"
          },
          "type": {
            "type": "string",
            "enum": ["subdomain", "custom", "wildcard"]
          },
          "target_ip": {
            "type": "string",
            "example": "127.0.0.1"
          },
          "port": {
            "type": "integer",
            "example": 3000
          }
        },
        "required": ["domain", "type"]
      },
...
    }
  },
  "paths": {
    "/api/domains": {
      "get": {
        "tags": ["Domains"],
        "summary": "List domains",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": ["Domains"],
        "summary": "Create domain",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDomainRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
...
    "/api/log": {
      "get": {
        "tags": ["Logs"],
        "summary": "Get logs",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": ["Logs"],
        "summary": "Clear logs",
        "responses": {
          "204": {
            "description": "Deleted"
          }
        }
      }
    },
...
  }
}